# internal/api/openapi.json
{
  "openapi": "3.2.1",
  "info": {
    "title": "WG-Guard API",
...
    "version": "1.0.0"
  },
  "servers": [{ "url": "/" }],
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "tags": ["users"], "summary": "Create a user", "operationId": "createUser",
        "description": "Idempotent with `Idempotency-Key`. When template_id names an enabled template, its technical terms override submitted limit, duration, start-policy and interface fields. Without a template, absent limits mean unlimited. This route creates no device; use /purchases for atomic customer-plus-first-device provisioning. Speed limits are independent per direction (kbps).",
        "security": [{ "BearerAuth": ["users.create"] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserCreate" }, "example": { "username": "alice", "duration_seconds": 2592000, "traffic_limit_bytes": 100000000000, "speed_limit_down_kbps": 100000, "speed_limit_up_kbps": null, "device_limit": 3 } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "400": { "$ref": "#/components/responses/Invalid" }, "409": { "$ref": "#/components/responses/Conflict" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      },
      "get": {
        "tags": ["users"], "summary": "List users (cursor pagination)", "security": [{ "BearerAuth": ["users.read"] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "maximum": 500, "default": 50 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["created_at", "username", "expires_at", "used"], "default": "created_at" } },
          { "name": "order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "username", "in": "query", "schema": { "type": "string" }, "description": "Substring filter (wildcards are literal)" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/UserStatus" } },
          { "name": "traffic_exceeded", "in": "query", "schema": { "type": "boolean" } },
          { "name": "enabled", "in": "query", "schema": { "type": "boolean" } },
          { "name": "expires_before", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "expires_after", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "created_before", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "created_after", "in": "query", "schema": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." } },
          { "name": "template_id", "in": "query", "schema": { "type": "string" } },
          { "name": "interface_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Page of users", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserPage" } } } } }
      }
    },
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": ["users"], "summary": "Get a user", "security": [{ "BearerAuth": ["users.read"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "responses": { "200": { "description": "User", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "patch": {
        "tags": ["users"], "summary": "Partially update a user (username immutable; `null` clears a limit)", "description": "Changing template_id on an existing user changes its reference only; it does not reset the existing entitlement. Update quota/duration fields explicitly or queue a successor.", "security": [{ "BearerAuth": ["users.update"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserPatch" } } } },
        "responses": { "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } }, "404": { "$ref": "#/components/responses/NotFound" } }
      },
      "delete": {
        "tags": ["users"], "summary": "Permanently delete an account and every device", "description": "One transaction removes the account, devices, customer link, chart history and queued next plan, releasing all its IP allocations. Username becomes available again. Former public peer keys remain as durable runtime removal intent until reconciliation succeeds. Audit and non-secret integration results retain their existing retention. Recovery requires a coordinated backup. Reseller tokens can delete only their own customers.", "security": [{ "BearerAuth": ["users.delete"] }],
        "parameters": [{ "$ref": "#/components/parameters/UserID" }],
        "responses": { "200": { "description": "Deleted" }, "404": { "$ref": "#/components/responses/NotFound" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "description": "API token: `Authorization: Bearer wg_…` with scopes; separate from admin sessions." }
    },
    "parameters": {
...
      "UserID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Unauthorized": { "description": "Missing or invalid token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "Token lacks the required scope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "Conflict (exists, limit reached, pool exhausted, in use)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Invalid": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "request_id"],
            "properties": {
              "code": { "type": "string", "description": "Stable machine code: USER_NOT_FOUND, USERNAME_EXISTS, DEVICE_LIMIT_REACHED, TRAFFIC_EXCEEDED, INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, RATE_LIMITED, NODE_UNAVAILABLE, INTERNAL_ERROR, …" },
              "message": { "type": "string" },
              "request_id": { "type": "string" }
            }
          }
        }
      },
...
      "UserStatus": { "type": "string", "enum": ["active", "disabled", "suspended", "expired", "traffic_exceeded", "waiting_first_connection"] },
      "StartPolicy": { "type": "string", "enum": ["immediate", "first_connection"], "description": "When the configured duration_seconds starts for a created subscription: immediate at creation, first_connection at the first observed connection. A queued Next Plan starts its duration at successor activation." },
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "username": {"type": "string", "description": "Immutable global username; use id in resource paths."},
          "display_name": {"type": "string"},
          "note": {"type": "string"},
          "tags": {"type": "array", "items": {"type": "string"}},
          "status": {"$ref": "#/components/schemas/UserStatus", "description": "Derived lifecycle state; not writable through user create/PATCH."},
          "disable_reason": {"type": ["string", "null"], "enum": ["manual", "expired", "traffic_limit", "admin_action", null]},
          "traffic_limit_bytes": {"type": ["integer", "null"], "description": "Quota in exact bytes (B). 1 GB = 1000000000 B; 100 GB = 100000000000 B. 100 GiB = 107374182400 B and displays as 107.4 GB. null = unlimited; 0 is a finite zero-byte allowance."},
          "traffic_used_rx": {"type": "integer", "description": "Charged received traffic in bytes (B), from client to server (upload)."},
          "traffic_used_tx": {"type": "integer", "description": "Charged sent traffic in bytes (B), from server to client (download)."},
          "traffic_used_total": {"type": "integer", "description": "Total charged traffic in bytes (B): traffic_used_rx + traffic_used_tx; both directions count toward quota."},
          "speed_limit_down_kbps": {"type": ["integer", "null"], "description": "Download (server to client) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 100000 = 100 Mbps = 12.5 MB/s in the panel speed form. null = unlimited."},
          "speed_limit_up_kbps": {"type": ["integer", "null"], "description": "Upload (client to server) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 20000 = 20 Mbps = 2.5 MB/s in the panel speed form. null = unlimited."},
          "device_limit": {"type": ["integer", "null"], "description": "Maximum registered configurations, including disabled devices; null is unlimited. It is not the number of configurations already provisioned."},
          "template_id": {"type": ["string", "null"], "description": "Technical template reference; existing terms are not automatically synchronized to later catalog edits."},
          "interface_id": {"type": ["string", "null"], "description": "User connection-profile reference; existing devices retain their own interface_id."},
          "start_policy": {"$ref": "#/components/schemas/StartPolicy"},
          "duration_seconds": {"type": ["integer", "null"], "description": "Duration in seconds: 86400 = 1 day; 2592000 = 30 days. The panel month shortcut means 30 days, not a calendar month. null = no configured duration."},
          "activated_at": {"type": ["string", "null"], "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds."},
          "expires_at": {"type": ["string", "null"], "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds."},
          "last_activity_at": {"type": ["string", "null"], "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds."},
          "enabled": {"type": "boolean"},
          "metadata": {"type": "object", "additionalProperties": true, "description": "Caller annotations stored at creation; current PATCH does not persist metadata updates."},
          "deleted": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds."},
          "updated_at": {"type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds."}
        }
      },
      "UserCreate": {
        "type": "object",
        "required": ["username"],
        "properties": {
          "username": {"type": "string", "pattern": "^[a-zA-Z0-9_-]{3,32}$", "description": "Globally unique immutable ASCII username, 3-32 letters/digits/underscore/hyphen. Resource paths require the returned opaque id, not this name. Permanent account deletion releases the name; a recreated account has a new opaque ID."},
          "display_name": {"type": "string", "description": "Optional presentation label distinct from the username. PATCH empty string clears it; absent/null leaves it unchanged."},
          "note": {"type": "string", "description": "Internal text; JSON \\n preserves line breaks. PATCH empty string clears it; absent/null leaves it unchanged. Never store credentials here."},
          "tags": {"type": "array", "items": {"type": "string"}, "description": "JSON string array, not comma-separated text. PATCH [] clears tags; absent/null leaves them unchanged."},
          "traffic_limit_bytes": {"type": ["integer", "null"], "description": "Quota in exact bytes (B). 1 GB = 1000000000 B; 100 GB = 100000000000 B. 100 GiB = 107374182400 B and displays as 107.4 GB. null = unlimited; 0 is a finite zero-byte allowance.", "minimum": 0},
          "speed_limit_down_kbps": {"type": ["integer", "null"], "description": "Download (server to client) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 100000 = 100 Mbps = 12.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1},
          "speed_limit_up_kbps": {"type": ["integer", "null"], "description": "Upload (client to server) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 20000 = 20 Mbps = 2.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1},
          "device_limit": {"type": ["integer", "null"], "description": "Maximum registered device/peer configuration count, including disabled devices. null = unlimited; zero is invalid. This does not allocate devices, detect physical hardware or count simultaneous connections. Lowering it does not delete existing devices. Use purchase device_count to provision a specific number of ready configurations.", "minimum": 1},
          "template_id": {"type": ["string", "null"], "description": "Optional enabled technical template ID. On create, saved quota, duration, start policy, device cap, speeds and interface override manual entitlement fields. Omit/null for manual terms."},
          "interface_id": {"type": ["string", "null"], "description": "Connection profile ID, not a name such as awg0. Omit/null for automatic selection when creating a device. A user PATCH does not move existing devices."},
          "start_policy": {"$ref": "#/components/schemas/StartPolicy"},
          "duration_seconds": {"type": ["integer", "null"], "description": "Duration in seconds: 86400 = 1 day; 2592000 = 30 days. The panel month shortcut means 30 days, not a calendar month. null = no configured duration."},
          "enabled": {"type": "boolean", "default": true, "description": "Account enabled flag (default true on create); status is a separate derived response field, not a writable request property."},
          "metadata": {"type": "object", "additionalProperties": true, "description": "Optional caller annotations saved on single-user creation. The current user PATCH handler accepts this property but does not persist updates; keep the editable billing journal outside WG-Guard."}
        },
        "additionalProperties": false
      },
      "UserPatch": {
        "type": "object",
        "properties": {
          "display_name": {"type": "string", "description": "Optional presentation label distinct from the username. PATCH empty string clears it; absent/null leaves it unchanged."},
          "note": {"type": "string", "description": "Internal text; JSON \\n preserves line breaks. PATCH empty string clears it; absent/null leaves it unchanged. Never store credentials here."},
          "tags": {"type": "array", "items": {"type": "string"}, "description": "JSON string array, not comma-separated text. PATCH [] clears tags; absent/null leaves them unchanged."},
          "traffic_limit_bytes": {"type": ["integer", "null"], "description": "Quota in exact bytes (B). 1 GB = 1000000000 B; 100 GB = 100000000000 B. 100 GiB = 107374182400 B and displays as 107.4 GB. null = unlimited; 0 is a finite zero-byte allowance.", "minimum": 0},
          "speed_limit_down_kbps": {"type": ["integer", "null"], "description": "Download (server to client) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 100000 = 100 Mbps = 12.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1},
          "speed_limit_up_kbps": {"type": ["integer", "null"], "description": "Upload (client to server) limit in decimal kilobits/s (1 Kbps = 1000 bits/s). 20000 = 20 Mbps = 2.5 MB/s in the panel speed form. null = unlimited.", "minimum": 1},
          "device_limit": {"type": ["integer", "null"], "description": "Maximum registered device/peer configuration count, including disabled devices. null = unlimited; zero is invalid. This does not allocate devices, detect physical hardware or count simultaneous connections. Lowering it does not delete existing devices. Use purchase device_count to provision a specific number of ready configurations.", "minimum": 1},
          "template_id": {"type": ["string", "null"], "description": "Optional technical template reference. PATCH only changes the reference and does not reapply saved terms or reset expiry/usage. null clears the reference."},
          "interface_id": {"type": ["string", "null"], "description": "Connection profile ID, not a name such as awg0. Omit/null for automatic selection when creating a device. A user PATCH does not move existing devices."},
          "duration_seconds": {"type": ["integer", "null"], "description": "Stored duration in seconds; positive values only. PATCH absent/null leaves it unchanged, and a value does not recalculate expires_at. Use renewal to change expiry. 2592000 = 30 days, the panel month shortcut."},
          "enabled": {"type": "boolean", "description": "Account enabled flag (default true on create); status is a separate derived response field, not a writable request property."},
          "metadata": {"type": "object", "additionalProperties": true, "description": "Optional caller annotations saved on single-user creation. The current user PATCH handler accepts this property but does not persist updates; keep the editable billing journal outside WG-Guard."}
        },
        "additionalProperties": false
      },
      "UserPage": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "next_cursor": { "type": "string", "description": "Opaque; empty = last page" }
        }
      },
...
    }
  },
...
}